Research Team/requirements.txt
semantic-kernel
openai
python-dotenv
